pyproject.toml
[project]
name = "impulseia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-unfold>=0.63.0",
    "djangorestframework>=3.16.0",
    "pillow>=11.3.0",
]
